+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.15.0] - 2020-01-01
+
+### Added
+- New module config setting for selecting indexed templates. This is essentially a shortcut for adding/removing index field to/from templates.
+
+### Changed
+- If $index_field_name for SearchEngine::getIndexField($index_field_name) is null, use default name from options.
+
 ## [0.14.0] - 2020-01-01
 
 ### Added